requirements.txt
lmstudio
fastmcp
gradio
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib